Added the new lighting infrastructure. Currently this only works for scene lighting, rather than point lighting.
This uses the OpenGL extension GL_ARB_framebuffer_object, but will fail gracefully if it's not detected.

The EnablePointSprite/DisablePointSprite functions have been removed - point sprites and scene lights now work the same way and use the same API to turn them on and off.

DrawLight, which is used for point sprites, is currently inoperable - I'll have this working soon.